Refactor account and user related views
