+++ monkeys/monkey_budget/views/__init__.py
@@ -1,6 +1,7 @@
 from .base import home, team, contact
-from .account import dashboard, show_money_account, add_money_account, edit_money_account, delete_money_account
+from .account import *
 from .transaction import (TransactionCreateView, TransactionUpdateView, TransactionDeleteView,
                           TransactionDetailView, TransactionListView,
                           AccountTransactionListView,
                           get_account_currency_info)
+from .user import *
